requirements.txt
flask
requests
APScheduler
python-dotenv
numpy==1.26.4
telegram
pandas
mplfinance
matplotlib
websocket-client



